+++ database/migrations/2022_01_26_094449_create_subscriptions_table.php
@@ -16,7 +16,7 @@ class CreateSubscriptionsTable extends Migration
         Schema::create('subscriptions', function (Blueprint $table) {
             $table->id();
 
-            $table->double('amount');
+            $table->decimal('amount');
             $table->timestamp('start_date');
             $table->timestamp('end_date');
             $table->string('status')->default('PENDING'); // PENDING / ACTIVE / EXPIRED
